+++ src/commands/interop/installer-module.ts
@@ -6,13 +6,13 @@ import type { MANAGER_JS } from "../../types/platform.ts";
 export const Installers = {
     UniJs: (path: string, manager: MANAGER_JS) => {
         Deno.chdir(ParsePath(path));
-        if (!ManagerExists(manager)) throw new FknError("Generic__MissingRuntime", `${manager} is not installed!`);
+        if (!ManagerExists(manager)) throw new FknError("Env__MissingMotor", `${manager} is not installed!`);
         Commander(manager, ["install"]);
         return;
     },
     Golang: (path: string) => {
         Deno.chdir(ParsePath(path));
-        if (!ManagerExists("go")) throw new FknError("Generic__MissingRuntime", `go is not installed!`);
+        if (!ManagerExists("go")) throw new FknError("Env__MissingMotor", `go is not installed!`);
         if (CheckForPath(JoinPaths(Deno.cwd(), "vendor"))) {
             Commander("go", ["mod", "vendor"]);
             return;
@@ -22,7 +22,7 @@ export const Installers = {
     },
     Cargo: (path: string) => {
         Deno.chdir(ParsePath(path));
-        if (!ManagerExists("cargo")) throw new FknError("Generic__MissingRuntime", `cargo is not installed!`);
+        if (!ManagerExists("cargo")) throw new FknError("Env__MissingMotor", `cargo is not installed!`);
         Commander("cargo", ["fetch"]);
         Commander("cargo", ["check"]);
         return;