+++ src/types/errors.ts
@@ -1,26 +1,61 @@
+export const OS_ERRORS = ["Os__NoAppdataNoHome"] as const;
+export const FS_ERRORS = ["Fs__Unreal", "Fs__UnparsablePath", "Fs__DemandsEmptying", "Fs__DemandsDIR"] as const;
+export const GIT_ERRORS = ["Git__UE", "Git__UE__Commit", "Git__UE_Push", "Git__UE__IsRepo", "Git__NoBranchAA", "Git__NoBranch"] as const;
+export const ENV_ERRORS = [
+    "Env__MissingMotor",
+    "Env__NoPkgFile",
+    "Env__PkgFileUnparsable",
+    "Env__CannotDetermine",
+    "Env__SchrodingerLockfile",
+] as const;
+export const TASK_ERRORS = ["Task__Release", "Task__Commit", "Task__Launch", "Task__Update", "Task__Lint", "Task__Pretty"] as const;
+export const PARAM_ERRORS = [
+    "Param__WhateverUnprovided",
+    "Param__TargetInvalid",
+    "Param__CIntensityInvalid",
+    "Param__SetupInvalid",
+    "Param__VerInvalid",
+    "Param__GitTargetInvalid",
+    "Param__GitTargetAliasInvalid",
+] as const;
+export const INTEROP_ERRORS = [
+    "Interop__PublishUnable",
+    "Interop__MigrateUnable",
+    "Interop__JSRunUnable",
+] as const;
+export const INTERNAL_ERRORS = [
+    "Internal__NonexistentAppPath",
+    "Internal__ImproperAssignment",
+    "Internal__InvalidEmbedded",
+    "Internal__Lazy",
+] as const;
+export const EXTERNAL_ERRORS = ["External__Publish", "External__Setting__FavIde", "External__Proj__NotFound"] as const;
+
+type OS_ERROR_CODES = typeof OS_ERRORS[number];
+type FS_ERROR_CODES = typeof FS_ERRORS[number];
+type GIT_ERROR_CODES = typeof GIT_ERRORS[number];
+type ENV_ERROR_CODES = typeof ENV_ERRORS[number];
+type TASK_ERROR_CODES = typeof TASK_ERRORS[number];
+type PARAM_ERROR_CODES = typeof PARAM_ERRORS[number];
+type INTEROP_ERROR_CODES = typeof INTEROP_ERRORS[number];
+type INTERNAL_ERROR_CODES = typeof INTERNAL_ERRORS[number];
+type EXTERNAL_ERROR_CODES = typeof EXTERNAL_ERRORS[number];
+
 /**
  * All possible CLI error codes.
  *
  * @export
  */
 export type GLOBAL_ERROR_CODES =
-    | "Cleaner__InvalidCleanerIntensity"
-    | "Internal__Projects__CantDetermineEnv"
-    | "Generic__NonExistingPath"
-    | "Internal__NoEnvForConfigPath"
-    | "Project__NonFoundProject"
-    | "Env__UnparsableMainFile"
-    | "Interop__CannotRunJsLike"
-    | "Unknown__CleanerTask__Lint"
-    | "Unknown__CleanerTask__Pretty"
-    | "Unknown__CleanerTask__Update"
-    | "Unknown__CleanerTask__Launch"
-    | "Commit__Fail__CommitCmd"
-    | "Sys__CannotNotify"
-    | "Git__IsRepoError"
-    | "Internal__UnparsablePath"
-    | "Generic__MissingRuntime"
-    | "Release__Fail__ReleaseCmd";
+    | OS_ERROR_CODES
+    | FS_ERROR_CODES
+    | GIT_ERROR_CODES
+    | ENV_ERROR_CODES
+    | TASK_ERROR_CODES
+    | PARAM_ERROR_CODES
+    | INTEROP_ERROR_CODES
+    | INTERNAL_ERROR_CODES
+    | EXTERNAL_ERROR_CODES;
 
 /**
  * All possible project validation error codes.